requirements.txt
# Core Framework and API
pydantic-ai==0.0.14
pydantic==2.10.3
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data Processing
pandas==2.2.3
numpy==2.2.1
openpyxl==3.1.5

# RAG and Embeddings
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
tiktoken==0.8.0

# PDF Generation
reportlab==4.4.4
xhtml2pdf==0.2.17
Pillow==11.0.0

# API Clients
httpx==0.28.1
requests==2.32.3
python-dotenv==1.0.1
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
